+++ src/components/Col/interfaces.ts
@@ -1,10 +1,10 @@
 type sizeOptions = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
 
-export interface IColProps {
+export type IColProps = {
   xs?: sizeOptions;
   sm?: sizeOptions;
   md?: sizeOptions;
   lg?: sizeOptions;
   xl?: sizeOptions;
   xxl?: sizeOptions;
-}
+};